+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: cnails <cnails@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/03 18:01:53 by cnails            #+#    #+#             */
-/*   Updated: 2019/11/28 17:55:27 by cnails           ###   ########.fr       */
+/*   Updated: 2019/12/04 20:43:27 by cnails           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,7 +70,7 @@ char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
 char				**ft_strsplit(char const *s, char c);
-char				*ft_itoa(int n);
+char				*ft_itoa(long int n);
 void				ft_putchar(char c);
 void				ft_putstr(char const *s);
 void				ft_putendl(char const *s);
@@ -94,6 +94,6 @@ int					get_next_line(const int fd, char **line);
 char				*ft_dtoa(double f, int n);
 char				*ft_strset(char c, size_t n);
 int					ft_for(int start, int sum, int end);
-char				*ft_itoa_base(unsigned int nbr, int base, char a);
+char				*ft_itoa_base(long long int nbr, int base, char a);
 
 #endif
